frontend/requirements.txt
Pyrebase4
streamlit
